Refactoring all classes
